Allow Card to be unit tested by making it not parcelable
Parcelable is only available in Android integration tests, which
is annoying to require for everything. We do not need to write out
the card list in the state anymore, as the state is now fully stored
in the DB, as we sort by age and not randomly anymore. So we can
get rid of the parcelable cards altogether.